feat: more granularity for liquidation penalty (#516)
* feat: convert liq price without penalty to tick

* fix: avoid invalid ticks

* fix: relax requirement

* feat!: more granularity for liquidation penalty

It is now possible to set the liquidation penalty to an arbitrary number of ticks, regardless of the tick spacing.

BREAKING CHANGE: the liquidation penalty is now expressed in ticks and is stored as a uint24. This affects the `LiquidationPenaltyUpdated` event, functions `getLiquidationPenalty`, `setLiquidationPenalty`, `getLongPosition`, `getTickLiquidationPenalty`, and the `TickDat` struct

* feat(long): expose new functions

* revert: don't make functions public

* refactor: mark functions internal

* test: fix some tests

* test: fix test

* test: temp fix for test

* test: temp fix test

* test: temp fix test

* test: fix tests

* test: dynamic initial deposit in fixture

* test: fix some tests

* test: fix tests

* test: fix tests

* test: fix more tests

* test: fix test

* test: fix test

* test: final fix

* test(long): add tests for `_getTickFromLiqPriceWithoutPenalty`

* refactor: rename function

* test: add fuzzing test

* test: fuzzing

* test: natspec

* test: fix fuzzing test

* test: increase fuzzing range

* docs: clarify natspec

* test: add fuzzing test

* test: typo

+++ src/UsdnProtocol/UsdnProtocolFallback.sol
@@ -206,7 +206,7 @@ contract UsdnProtocolFallback is IUsdnProtocolFallback, UsdnProtocolStorage {
     }
 
     /// @inheritdoc IUsdnProtocolFallback
-    function getLiquidationPenalty() external view returns (uint8) {
+    function getLiquidationPenalty() external view returns (uint24) {
         return s._liquidationPenalty;
     }
 
@@ -517,7 +517,7 @@ contract UsdnProtocolFallback is IUsdnProtocolFallback, UsdnProtocolStorage {
     }
 
     /// @inheritdoc IUsdnProtocolFallback
-    function setLiquidationPenalty(uint8 newLiquidationPenalty) external onlyRole(SET_PROTOCOL_PARAMS_ROLE) {
+    function setLiquidationPenalty(uint24 newLiquidationPenalty) external onlyRole(SET_PROTOCOL_PARAMS_ROLE) {
         if (newLiquidationPenalty > Constants.MAX_LIQUIDATION_PENALTY) {
             revert IUsdnProtocolErrors.UsdnProtocolInvalidLiquidationPenalty();
         }
